#6 delivery method (#134)
* style the setiings tabs and move pakjegemak

* style the setings tabs and add shippingmethods

* settings tabs

* Set pakjegemak config

* new MyParcel checkout template

* Shipping methods

* set template

* chance url for request

* setup request

* api2 request error

* api2 send data in createConsignmentRequest

* api2 save consignment ID

* merge develop from 1.4.0-extra 2

* fix conflict

* log new api

* log new api

* new api

* generate pdf

* generate and download pdf

* Generating positions for A4 paper

* Generating positions for A4 paper

* fix request int

* fix request int

* no eps_postal_code for pakjegemak

* pakjegemak API 2

* show error

* show error

* check config for GET

* get positions clear

* trim json data + set api key

* clean code

* cd shipments

* cd shipments contents

* HS code

* Error; No Customs Content HS Code found

* Error; International shipments can not be sent by letter box

* Error; no HS code

* Total weight * 1000

* Total value * 100

* update status after create consignment

* update status after create consignment

* show order status in shipment grid

* update @var

* sales shipment after load ->status update

* phone number

* cronjob update status

* cleanup

* update from 1.4.1

* update cronjob

* update cronjob

* bolcom fix

* update upgrade

* Connect MyParcelShipment with track

* Set good version

* track set back

* track set back

* Add column is_xl to database

* save is_xl to shipment

* set default

* send large format

* translate large package

* translate large package

* Update from 1.5.0-beta

* delete mp_checkout

* large package

* ignore

* new checkout files

* new MyParcel options in checkout

* edit url and css

* get Magento url

* set js controller

* set js controller deliveryType

* 1.9

* 0.9

* $ jQuery conflict

* save shipmentmethods

* checkout save shipping method

* gitignore

* base url

* $ conflict

* show options

* move saveMyParcelShippingMethod and save address

* comment

* comment

* connect myparcel data quote tot order

* iwd ignore

* set order datetime

* translate bezorgen

* delivery_date

* delivery_type

* show html after load

* options view ' '

* update simon

* update simon

* update simon

* more clear

* save return if no answer

* save return if no answer

* data from object tot array

* data from object tot array

* data from object tot array

* to array

* translation

* get home selected

* set 1.6.0 ipv 1.6.1

* set 1.6.0 ipv 1.6.1

* myparcel_date sql

* morging

* no free

* settings

* no pakjegemak

* retail

* set drop-off settings

* set drop-off settings

* Checkout get fullstreet

* Checkout regex

* Checkout get house number

* Checkout get house number

* checkout settings

* checkout settings

* checkout settings

* checkout settings (days)

* checkout settings

* get checkout data

* drop off days setting

* settings

* mp checkout settings

* config

* checkout  v1.0.1

* checkout

* checkout

* signed

* base price

* base price

* _getDropOffDay

* show Today

* show Today

* show Today

* tmp disable min_order

* checkout and show select color

* js color

* text and color

* colors

* pg go back

* pg go back

* set price

* set price

* update sql version

* set version 1.6.1

* tmp remove for fix

* revert for uppercase

* sql data type

* set home address only

* checkout

* update delivery

* getInfo -> info

* Get base url

* set ipv get

* mypa load

* update country status

* move delivery

* show error

* check if mypa is set

* clean up

* clean up

* clean up

* calculate price

* Set checkout shipment price

* Checkout 'Huisnummer is niet aanwezig'

* Translation

* Translations

* set iwd opc_index_index.js

* support iwd checkout

* support iwd checkout

* support iwd checkout

* support iwd checkout

* support iwd checkout

* support iwd checkout

* support iwd checkout + delete MutationObserver

* support iwd checkout

* support iwd checkout

* support iwd checkout

* no !important;

* no !important;

* create return

* create return

* clean up

* clean up

* fix street checkout

* Fix direct select 'Niet bij de buren'

* dropoff_days

* Rewrite available.phtml

* remove console.log

* checkout

* checkout

* always_myparcel

* always_myparcel + support multi address

* checkout

* checkout max height

* checkout max height + icon

* fix checkout checked

* fix checkout checked

* Merge branches '#6-delivery-method' and 'develop' of https://github.com/myparcelnl/magento1 into #6-delivery-method

# Conflicts:
#	.gitignore
#	app/code/community/TIG/MyParcel2014/Block/Adminhtml/Sales/Order/View/ShippingInfo.php
#	app/code/community/TIG/MyParcel2014/Block/Adminhtml/Widget/Grid/Column/Renderer/ShippingStatus.php
#	app/code/community/TIG/MyParcel2014/Helper/Data.php
#	app/code/community/TIG/MyParcel2014/Model/Api/MyParcel.php
#	app/code/community/TIG/MyParcel2014/Model/Carrier/MyParcel.php
#	app/code/community/TIG/MyParcel2014/Model/Checkout/Service.php
#	app/code/community/TIG/MyParcel2014/Model/Observer/Cron.php
#	app/code/community/TIG/MyParcel2014/Model/Sales/Order.php
#	app/code/community/TIG/MyParcel2014/etc/config.xml
#	app/code/community/TIG/MyParcel2014/etc/system.xml
#	app/design/adminhtml/default/default/template/TIG/MyParcel2014/sales/order/mass_actions.phtml
#	app/design/adminhtml/default/default/template/TIG/MyParcel2014/sales/order/shipment/shipping_info.phtml
#	app/design/adminhtml/default/default/template/TIG/MyParcel2014/sales/order/view/shipping_info.phtml
#	app/design/frontend/base/default/layout/TIG/tig_myparcel.xml
#	app/design/frontend/base/default/template/TIG/MyParcel2014/checkout/idev_onestepcheckout/js.phtml
#	app/locale/nl_NL/TIG_MyParcel2014.csv
#	skin/frontend/base/default/js/TIG/MyParcel2014/checkout.js

* xl for eu

* checkout use billing/ shipping

* translation Dropoff-delay

* default xl shipment

* add shipment track and trace

* Status in detailpagina

* base_url

* customs_declaration contents

* status

* retour mail NL only

* unpaid for all countries

* set staging url

* fix label position a4

* space in zip

* no default postal code

* no default postal code

* fix checkbox not NL and load 1 time

* show version

* Checkout improvements

* signature_and_only_recipient

* checkout with template

* checkout with template

* checkout with template

* implement <template checkout

* implement <template checkout

* iwd checkout

* jquery load

* move colors

* move colors

* move colors

* ontstepcheckout

* ontstepcheckout

* get address from quote for the ajax request

* ontstepcheckout submit

* ignore

* pull from develop

* pull from develop

* 1.6.1

+++ app/code/community/TIG/MyParcel2014/Model/Checkout/Service.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  *                  ___________       __            __
  *                  \__    ___/____ _/  |_ _____   |  |
@@ -39,12 +40,75 @@
 class TIG_MyParcel2014_Model_Checkout_Service
 {
     /**
-     * @param array                  $addressData
+     * Save in checkout MyParcel shipping method
+     */
+    public function saveMyParcelShippingMethod()
+    {
+        $request = Mage::app()->getRequest();
+        if($request->isPost()){
+
+            $helper = Mage::helper('tig_myparcel');
+            $addressHelper = Mage::helper('tig_myparcel/addressValidation');
+            $quote = Mage::getModel('checkout/cart')->getQuote();
+
+            $address = $addressHelper->getQuoteAddress($quote);
+
+            if ($address['country'] !== 'NL')
+                return true;
+
+            /**
+             * If shipping method is myparcel
+             */
+            if (strpos($request->getPost('shipping_method', ''), 'myparcel') !== false) {
+
+
+                if ($request->getPost('mypa-input') == null)
+                    return false;
+
+                $data = json_decode($request->getPost('mypa-input', ''), true);
+
+                if ($data['location'] !== null) {
+
+                    /**
+                     * is pickup
+                     */
+                    $this->savePgAddress($data, $quote);
+                } else {
+                    /**
+                     * not pickup
+                     */
+                    $return = $request->getPost('mypa-recipient-only', '') === 'on' ? 1 : false;
+                    if ($return) {
+                        $data['home_address_only'] = true;
+                    }
+
+                    $signed = $request->getPost('mypa-signed', '') === 'on' ? 1 : false;
+                    if ($signed) {
+                        $data['signed'] = true;
+                    }
+
+
+                    $this->removePgAddress($quote);
+                }
+
+                $quote->setMyparcelData(json_encode($data))->save();
+                $helper->updateRatePrice($quote);
+
+            } else {
+                $quote->setMyparcelData(null)->save();
+                $this->removePgAddress(json_encode($quote));
+            }
+        }
+        return true;
+    }
+
+    /**
+     * @param object                 $data
      * @param Mage_Sales_Model_Quote $quote
      *
      * @return $this
      */
-    public function savePgAddress(array $addressData, Mage_Sales_Model_Quote $quote)
+    public function savePgAddress($data, Mage_Sales_Model_Quote $quote)
     {
         $helper = Mage::helper('tig_myparcel');
 
@@ -58,20 +122,20 @@ class TIG_MyParcel2014_Model_Checkout_Service
          */
         $pgAddress = Mage::getModel('sales/quote_address');
         $pgAddress->setAddressType($helper::PG_ADDRESS_TYPE)
-                  ->setCity($addressData['city'])
-                  ->setCountryId('NL')
-                  ->setPostcode($addressData['postcode'])
-                  ->setCompany($addressData['name'])
-                  ->setFirstname('Ophalen op een PostNL locatie')
-                  ->setLastname('')
-                  ->setTelephone($addressData['telephone'])
-                  ->setStreet($addressData['street'] . "\n" . $addressData['housenr']);
+            ->setCity($data['city'])
+            ->setCountryId('NL')
+            ->setPostcode($data['postal_code'])
+            ->setCompany($data['location'])
+            ->setFirstname('Ophalen op een PostNL locatie')
+            ->setLastname('')
+            ->setTelephone($data['phone_number'])
+            ->setStreet($data['street'] . "\n" . $data['number']);
 
         /**
          * Add the address to the quote and save the quote.
          */
         $quote->addAddress($pgAddress)
-              ->save();
+            ->save();
 
         /**
          * Save the address.
@@ -120,7 +184,7 @@ class TIG_MyParcel2014_Model_Checkout_Service
         $orderAddress = Mage::getModel('sales/convert_quote')->addressToOrderAddress($address);
 
         $order->addAddress($orderAddress)
-              ->save();
+            ->save();
 
         /**
          * This is a fix for the order address missing a parent ID.
@@ -147,12 +211,16 @@ class TIG_MyParcel2014_Model_Checkout_Service
      */
     public function removePgAddress(Mage_Sales_Model_Quote $quote)
     {
-        $addresses = $quote->getAllAddresses();
+        if ($quote) {
+            if ($quote != '{}') {
+                $addresses = $quote->getAllAddresses();
 
-        /** @var Mage_Sales_Model_Quote_Address $address */
-        foreach ($addresses as $address) {
-            if ($address->getAddressType() == TIG_MyParcel2014_Helper_Data::PG_ADDRESS_TYPE) {
-                $address->delete();
+                /** @var Mage_Sales_Model_Quote_Address $address */
+                foreach ($addresses as $address) {
+                    if ($address->getAddressType() == TIG_MyParcel2014_Helper_Data::PG_ADDRESS_TYPE) {
+                        $address->delete();
+                    }
+                }
             }
         }
